+++ druid/src/lens/lens.rs
@@ -575,3 +575,29 @@ where
         v
     }
 }
+
+/// A `Lens` that always yields ().
+///
+/// This is useful when you wish to have a display only widget, require a type-erased widget, or
+/// obtain app data out of band and ignore your input. (E.g sub-windows)
+#[derive(Debug, Copy, Clone)]
+pub struct Unit<T> {
+    phantom_t: PhantomData<T>,
+}
+
+impl<T> Default for Unit<T> {
+    fn default() -> Self {
+        Unit {
+            phantom_t: Default::default(),
+        }
+    }
+}
+
+impl<T> Lens<T, ()> for Unit<T> {
+    fn with<V, F: FnOnce(&()) -> V>(&self, _data: &T, f: F) -> V {
+        f(&())
+    }
+    fn with_mut<V, F: FnOnce(&mut ()) -> V>(&self, _data: &mut T, f: F) -> V {
+        f(&mut ())
+    }
+}